更新商品增加级联各类接口

+++ Coupon/src/main/java/com/newhopemail/coupon/entity/SkuLadderEntity.java
@@ -32,7 +32,7 @@ public class SkuLadderEntity implements Serializable {
 	/**
 	 * 满几件
 	 */
-	private Integer fullCount;
+	private Long fullCount;
 	/**
 	 * 打几折
 	 */
